requirements.txt
boto==2.13.3
distribute==0.7.3
dj-cmd==0.4
dj-database-url==0.2.2
django-appconf==0.6
django-bootstrap-form==3.0
django-countries==1.5
django-filter==0.7
django-heroku-memcacheify==0.4
django-heroku-postgresify==0.3
django-mptt==0.6.0
django-pylibmc-sasl==0.2.4
django-reset==0.2.0
django-storages==1.1.8
django-suit==0.2.5
Django==1.5.4
djangorestframework==2.3.8
gevent==0.13.8
greenlet==0.4.1
gunicorn==18.0
ipython==1.1.0
Pillow==2.2.1
pip-tools==0.3.4
psycopg2==2.5.1
pylibmc==1.2.3
pytz==2013.7
raven==3.5.0
six==1.4.1
sorl-thumbnail==11.12
South==0.8.2
